More file position data type changes. Also reflecting this in user code.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@954 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/COBJMeshFileLoader.cpp
@@ -61,7 +61,7 @@ bool COBJMeshFileLoader::isALoadableFileExtension(const c8* filename)
 //! See IReferenceCounted::drop() for more information.
 IAnimatedMesh* COBJMeshFileLoader::createMesh(io::IReadFile* file)
 {
-	const u32 filesize = file->getSize();
+	const long filesize = file->getSize();
 	if (!filesize)
 		return 0;
 
@@ -288,7 +288,7 @@ void COBJMeshFileLoader::readMTL(const c8* pFileName, core::stringc relPath)
 	if (!pMtlReader)	// fail to open and read file
 		return;
 
-	const u32 filesize = pMtlReader->getSize();
+	const long filesize = pMtlReader->getSize();
 	if (!filesize)
 		return;
